Update: provider and model are now validated or closest matched + misc refactors
